Fix family members loading by improving getFamilyMembers() error handling
- Remove artificial 100-user limit in fallback queries to ensure all family members are found
- This fixes the 'unable to load family members' error in Recurring Payments screen
- The method now fetches all users when indexed queries fail, ensuring complete family member discovery

+++ lib/widgets/budget/spending_trends_chart.dart
@@ -60,7 +60,7 @@ class SpendingTrendsChart extends StatelessWidget {
               sideTitles: SideTitles(
                 showTitles: true,
                 reservedSize: 30,
-                interval: sortedDates.length > 7 ? (sortedDates.length / 7).ceil() : 1,
+                interval: sortedDates.length > 7 ? (sortedDates.length / 7).ceil().toDouble() : 1.0,
                 getTitlesWidget: (value, meta) {
                   final index = value.toInt();
                   if (index >= 0 && index < sortedDates.length) {
